Added ability to read compute messages
I added the functionality to read the compute
messages from the HEC-RAS. Every time the program
runs it spits out a <model plan>.computeMsgs.txt
text file that contains all of the information from
the Computation Messages screen. This way, the user
can check that file for errors if they occurred
(it gives specific information such as problematic xs)
or other things of interest (such as if if the model
did not converge because the model will run without errors
in that case).